+++ src/DateTimeOffset.cpp
@@ -31,13 +31,14 @@
  *          timezone-aware datetime values with 100-nanosecond precision.
  */
 
+#include "nfx/datetime/DateTimeOffset.h"
+#include "Internal.h"
+
+#include <nfx/string/StringBuilder.h>
+
 #include <charconv>
 #include <istream>
 #include <stdexcept>
-#include <sstream>
-
-#include "nfx/datetime/DateTimeOffset.h"
-#include "Internal.h"
 
 namespace nfx::time
 {